requirements.txt
absl-py==0.11.0
apturl==0.5.2
astunparse==1.6.3
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.1
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.0
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
GeoIP==1.3.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hkdf==0.0.3
html5lib==1.0.1
httplib2==0.14.0
humanize==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
magic-wormhole==0.11.2
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
notebook==6.0.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pandas==1.2.1
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.14.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.15.0
spake2==0.8
systemd-python==234
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
tornado==5.1.1
tqdm==4.30.0
traitlets==4.3.3
Twisted==18.9.0
txaio==2.10.0
txtorcon==19.0.0
typing-extensions==3.7.4.3
u-msgpack-python==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==2.0.0
wrapt==1.12.1
wsaccel==0.6.2
xkit==0.0.0
xport==3.2.1
zipp==1.0.0
zope.interface==4.7.1
